Switch SecretsSettingsSource to the env variable parsing logic
SecretsSettingsSource now inherits from EnvSettingsSourceBase and reuses the
parsing logic for complex and nested variables available for env variables.

This patch slightly changes the overall behavior with regard to reading secrets
files: In the old implementation, secret file paths were built on-demand from
the configuration fields, and the code then attempted to open the secret file.
In the new implementation, the code reads all available secrets in batch before
attempting to match them to fields. In order to avoid potential permission
problems if there are secrets for different apps in the same container / pod,
it is no longer an error if the secret file cannot be read - the code issues a
warning instead.